Changed return type of Packet.GetShort
It's very unlikely we would ever want an actual short, so we might as well return an int. Makes our life easier.

+++ src/Shared/Network/Packet.cs
@@ -161,7 +161,7 @@ namespace Sabine.Shared.Network
 		/// Reads a short from packet and returns it.
 		/// </summary>
 		/// <returns></returns>
-		public short GetShort()
+		public int GetShort()
 			=> _buffer.ReadInt16();
 
 		/// <summary>
